python/tutorials/solitaire/solitaire-final-async/requirements.txt
flet-pyodide
